辛勤劳动的我，已经不知道改了什么了

+++ src/main/java/com/yunquanlai/admin/product/service/ProductInfoService.java
@@ -26,9 +26,9 @@ public interface ProductInfoService {
 	
 	void update(ProductInfoVO productInfoVO);
 	
-	void delete(Integer id);
+	void delete(Long id);
 	
-	void deleteBatch(Integer[] ids);
+	void deleteBatch(Long[] ids);
 
 	List<ProductInfoEntity> queryListForClient(Map<String, Object> map);
 }
